+++ RELEASE NOTES.md
@@ -1,3 +1,8 @@
+Version 1.1
+
+- Added animated image sequence playback to GLImageView
+- Added PVR video demo to the example project
+
 Version 1.0
 
 - Initial release